Fix build errors: missing includes, format specifiers, snprintf buffer
- display.c, touch.c: add missing esp_timer.h include
- main.c: add lvgl.h include for lv_timer_handler
- encoder.c: cast int32_t to int for %d format specifier
- web_server.c: increase snprintf buffer, fix misleading indentation,
  shorten label text to fit within buffer
- CMakeLists.txt: remove mdns from REQUIRES (moved to component manager)
- idf_component.yml: add espressif/mdns as managed component

+++ main/encoder.c
@@ -84,6 +84,6 @@ esp_err_t encoder_init(void) {
     ESP_ERROR_CHECK(esp_timer_create(&args, &timer));
     ESP_ERROR_CHECK(esp_timer_start_periodic(timer, 3000));
 
-    ESP_LOGI(TAG, "encoder ready, vol_step=%d (%.1f dB)", s_vol_step, s_vol_step / 10.0f);
+    ESP_LOGI(TAG, "encoder ready, vol_step=%d (%.1f dB)", (int)s_vol_step, s_vol_step / 10.0f);
     return ESP_OK;
 }
